requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
Django==2.2.12
django-cors-headers==3.3.0
django-rest-swagger==2.2.0
django-silk==4.0.1
djangorestframework==3.9.2
et-xmlfile==1.0.1
gprof2dot==2019.11.30
idna==2.9
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.4
openapi-codec==1.3.2
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.7
Pygments==2.6.1
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
pytz==2020.1
qrcode==6.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
